requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.5.0
eventlet==0.35.1
Flask==3.0.2
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
Werkzeug==3.0.1
